requirements.txt
fastapi
uvicorn[standard]
websockets
groq
openai-whisper
gTTS
pyttsx3
langdetect
fasttext-wheel
redis
aioredis
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
celery[redis]
structlog
python-dotenv
pytest
pytest-asyncio
httpx
python-multipart
